Explicitly mention CC license versions
This means that if the owner of a problem package has licensed his
problem hypothetically in the future under CC by-sa 5, then he can
still select `cc -by-sa` here without the specification implying
that it would actually be licensed under version 4.

For future versions of the specification we can bump the CC license
version to the latest version (if any new version was released),
assuming that we want to require problem owners to always license
under the latest version of a CC license.

Also backport this to the legacy specification.